requirements.txt
python-dotenv
pydantic
openai
requests
bs4
networkx
pyvis
fastapi
uvicorn